Avoid -Wsign-conversion warning in detail/basic_name_generator.hpp when wchar_t is int32_t

+++ include/boost/uuid/detail/basic_name_generator.hpp
@@ -87,9 +87,9 @@ private:
     {
         BOOST_UUID_STATIC_ASSERT( sizeof( std::uint32_t ) >= sizeof( wchar_t ) );
 
-        for( std::size_t i = 0; i < n; ++i)
+        for( std::size_t i = 0; i < n; ++i )
         {
-            std::uint32_t ch = p[ i ];
+            std::uint32_t ch = static_cast<std::uint32_t>( p[ i ] );
 
             unsigned char bytes[ 4 ] =
             {
